[#1426] moved mapview to offlinemaps module
