+++ app/src/main/res/layout/activity_data_point_map.xml
@@ -8,7 +8,7 @@
 
     <include layout="@layout/app_bar_layout"/>
 
-    <org.akvo.flow.presentation.datapoints.map.one.MapBoxMapItemView
+    <org.akvo.flow.offlinemaps.presentation.MapBoxMapItemView
         android:id="@+id/mapView"
         android:layout_marginTop="?actionBarSize"
         android:layout_width="match_parent"
